Updated version dependencies
The version listed in the dependencie will not work if one has a newer version of Clojure - even if the description says version 1.5 or up the program won't run with version 1.10.0 for example. If you are learning Clojure with Clojure for the Brave and True then you will have the latest version installed, changing the dependency here makes it possible to only need one verison of Coljure installed in order to run both the projects in the book as well as the clojure-katas.